Rename src/image/index.js -> src/image/image.js
